+++ src/spring/azext_spring/tests/latest/test_asa_scenario.py
@@ -228,3 +228,15 @@ class CustomImageTest(ScenarioTest):
             self.check('properties.source.type', 'Container'),
             self.check('properties.source.customContainer.containerImage', '{containerImage}'),
         ])
+
+class AppConnectTest(ScenarioTest):
+
+    def test_app_connect(self):
+        self.kwargs.update({
+            'app': 'test-app',
+            'serviceName': 'cli-unittest',
+            'resourceGroup': 'cli'
+        })
+
+        # Test the failed case only since this is an interactive command
+        self.cmd('spring app connect -s {serviceName} -g {resourceGroup} -n {app} --shell-cmd /bin/placeholder', expect_failure=True)